x/blue/types/genesis_test.go
package types_test

import (
	"testing"

	"blue/x/blue/types"
	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	for _, tc := range []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				PrefPoolList: []types.PrefPool{
					{
						Id: 0,
					},
					{
						Id: 1,
					},
				},
				PrefPoolCount: 2,
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated prefPool",
			genState: &types.GenesisState{
				PrefPoolList: []types.PrefPool{
					{
						Id: 0,
					},
					{
						Id: 0,
					},
				},
			},
			valid: false,
		},
		{
			desc: "invalid prefPool count",
			genState: &types.GenesisState{
				PrefPoolList: []types.PrefPool{
					{
						Id: 1,
					},
				},
				PrefPoolCount: 0,
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	} {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
